# Improve CFB8 implementation