--- a/net/CFB8/cfb8.go
+++ b/net/CFB8/cfb8.go
@@ -6,46 +6,59 @@ import "crypto/cipher"
 type CFB8 struct {
 	c         cipher.Block
 	blockSize int
-	iv, tmp   []byte
+	ivPos     int
+	iv        []byte
 	de        bool
 }
 
 func NewCFB8Decrypt(c cipher.Block, iv []byte) *CFB8 {
-	cp := make([]byte, len(iv))
+	cp := make([]byte, len(iv)*3)
 	copy(cp, iv)
 	return &CFB8{
 		c:         c,
 		blockSize: c.BlockSize(),
 		iv:        cp,
-		tmp:       make([]byte, c.BlockSize()),
 		de:        true,
 	}
 }
 
 func NewCFB8Encrypt(c cipher.Block, iv []byte) *CFB8 {
-	cp := make([]byte, len(iv))
+	cp := make([]byte, len(iv)*3)
 	copy(cp, iv)
 	return &CFB8{
 		c:         c,
 		blockSize: c.BlockSize(),
 		iv:        cp,
-		tmp:       make([]byte, c.BlockSize()),
 		de:        false,
 	}
 }
 
 func (cf *CFB8) XORKeyStream(dst, src []byte) {
 	for i := 0; i < len(src); i++ {
+		posPlusBlockSize := cf.ivPos + cf.blockSize
 		val := src[i]
-		copy(cf.tmp, cf.iv)
-		cf.c.Encrypt(cf.iv, cf.iv)
-		val = val ^ cf.iv[0]
+		// fast mod; 2*blockSize must be a non-negative integer power of 2
+		tempPos := posPlusBlockSize & (cf.blockSize<<1 - 1)
+		cf.c.Encrypt(cf.iv[tempPos:], cf.iv[cf.ivPos:])
+		val ^= cf.iv[tempPos]
 
-		copy(cf.iv, cf.tmp[1:])
-		if cf.de {
-			cf.iv[15] = src[i]
+		if cf.ivPos > cf.blockSize {
+			if cf.de {
+				cf.iv[cf.ivPos-cf.blockSize-1] = src[i]
+			} else {
+				cf.iv[cf.ivPos-cf.blockSize-1] = val
+			}
+		}
+
+		if cf.ivPos == cf.blockSize<<1 {
+			cf.ivPos = 0
 		} else {
-			cf.iv[15] = val
+			if cf.de {
+				cf.iv[posPlusBlockSize] = src[i]
+			} else {
+				cf.iv[posPlusBlockSize] = val
+			}
+			cf.ivPos += 1
 		}
 
 		dst[i] = val