Before:

```
goos: windows
goarch: amd64
pkg: github.com/Tnze/go-mc/net/CFB8
cpu: Intel(R) Core(TM) i5-10210U CPU @ 1.60GHz
BenchmarkCFB8AES1K
BenchmarkCFB8AES1K-8       52201             23038 ns/op          44.45 MB/s               0 B/op          0 allocs/op
PASS
```

After:

```
goos: windows
goarch: amd64
pkg: github.com/Tnze/go-mc/net/CFB8
cpu: Intel(R) Core(TM) i5-10210U CPU @ 1.60GHz
BenchmarkCFB8AES1K
BenchmarkCFB8AES1K-8       52834             21968 ns/op          46.61 MB/s               0 B/op          0 allocs/op
PASS
```

Thanks @lycorisor for benchmark result on his laptop.