@layou233 2023-06-23
Also tested with:
```go
// License: MIT
package main

import (
	"crypto/aes"
	"encoding/base64"
	"encoding/hex"
	"fmt"
	
	"github.com/Tnze/go-mc/net/CFB8"
)

func main() {
	cipher, _ := hex.DecodeString("2b7e151628aed2a6abf7158809cf4f3c")
	iv, _ := hex.DecodeString("000102030405060708090a0b0c0d0e0f")
	block, _ := aes.NewCipher(cipher)
	stream := CFB8.NewCFB8Encrypt(block, iv) // or CFB8.NewCFB8Decrypt
	for {
		s := ""
		fmt.Scanln(&s)
		text := []byte(s)
		stream.XORKeyStream(text, text)
		fmt.Println(base64.StdEncoding.EncodeToString(text))
	}
}
```
The output of the new implementation matches the old one (https://github.com/Tnze/go-mc/blob/fad92fe3640c7dcfff04d90632b08dedfeaa3418/net/CFB8/cfb8.go).

I think it is safe to merge.