@layou233 2023-06-24
Now faster:

```
goos: windows
goarch: amd64
pkg: github.com/Tnze/go-mc/net/CFB8
cpu: Intel(R) Xeon(R) Platinum 8272CL CPU @ 2.60GHz
BenchmarkCFB8AES1K
BenchmarkCFB8AES1K-8       55309             21203 ns/op          48.30 MB/s               0 B/op          0 allocs/op
PASS
```