@layou233 2023-06-25
Faster again when dst and src are not overlap:
```
cpu: Intel(R) Xeon(R) Platinum 8272CL CPU @ 2.60GHz
BenchmarkCFB8AES1KOverlap
BenchmarkCFB8AES1KOverlap-8                55369             21358 ns/op          47.95 MB/s           0 B/op          0 allocs/op
BenchmarkCFB8AES1KNonOverlap
BenchmarkCFB8AES1KNonOverlap-8             61537             19498 ns/op          52.52 MB/s           0 B/op          0 allocs/op
PASS
```