@layou233 2023-06-27
> Does it seem like we should upgrade to Go 1.20?

Go 1.21 has not been released yet so Go 1.19 is still under maintenance. It may not the time to drop support for Go 1.19.

A simple make and copy is good enough for this, and I did that change.

## Yet another summary for this PR

1. Fix: The original implementation assumed a block size of 16, breaking support with other block cipher algorithms (such as DES).
2. Improve: Less make, less copy. We copy 15 additional bytes only after processing 32 bytes, while the original implementation copies 15 additional bytes every operation. If the length of `src` is greater than 2*block_size and the `dst` and `src` are not overlapping*, we'll use an optimized implementation that copy only 15 additional bytes during the entire process.
3. Test: Add tests for CFB8, including overlapped/non-overlapping operation test, multi-call test and benchmark.

*: If `dst` and `src` are inexactly overlapping and the pointer (location in memory) to first byte of `dst` is more than 16 bytes (or equal) ahead of the pointer to first byte of `src`, they are still considered "not overlapping" in this implementation and would receive optimization.