@layou233 2023-07-09
Tested with a simple bot:
```go
func main() {
	b := bot.NewClient()
	b.Auth = bot.Auth{
		Name: "la_you",
		UUID: "fd38a9a2-d42d-448e-bd8a-10f802ee9127",
		AsTk: "some access token, removed :-)", // TODO: fill access token
	}
	b.Name = "la_you"
	b.UUID = uuid.MustParse("fd38a9a2-d42d-448e-bd8a-10f802ee9127")

	err := b.JoinServer(":25565") // hosted on my local machine
	if err != nil {
		panic(err)
	}

	playerList := playerlist.New(b)
	msg.New(b,
		basic.NewPlayer(b, basic.DefaultSettings, basic.EventsListener{}),
		playerList,
		msg.EventsHandler{
			SystemChat: func(msg chat.Message, overlay bool) error {
				fmt.Println("SystemChat:", msg.String(), overlay)
				return nil
			},
			PlayerChatMessage: func(msg chat.Message, validated bool) error {
				fmt.Println("PlayerChat:", msg.String(), validated)
				return nil
			},
			DisguisedChat: func(msg chat.Message) error {
				fmt.Println("DisguisedChat:", msg.String())
				return nil
			},
		},
	)

	panic(b.HandleGame())
}
```
Test server: Mojang vanilla Minecraft server, version 1.19.4, with `online-mode=true`.

And it works as expected:
![image](https://github.com/Tnze/go-mc/assets/23739827/d1f0eaca-29cc-4511-b9ac-487661533e40)